NODE-550: Move 'toKey' function to the separate 'Utils' module
