+++ node/src/main/scala/io/casperlabs/node/api/GrpcCasperService.scala
@@ -2,8 +2,8 @@ package io.casperlabs.node.api
 
 import cats.effect._
 import cats.implicits._
-import com.google.protobuf.empty.Empty
 import com.google.protobuf.ByteString
+import com.google.protobuf.empty.Empty
 import io.casperlabs.blockstorage.BlockStore
 import io.casperlabs.casper.MultiParentCasperRef.MultiParentCasperRef
 import io.casperlabs.casper.SafetyOracle
@@ -12,14 +12,13 @@ import io.casperlabs.casper.consensus.Block
 import io.casperlabs.casper.consensus.info._
 import io.casperlabs.casper.consensus.state
 import io.casperlabs.catscontrib.MonadThrowable
+import io.casperlabs.comm.ServiceError.InvalidArgument
+import io.casperlabs.ipc
 import io.casperlabs.metrics.Metrics
+import io.casperlabs.models.SmartContractEngineError
 import io.casperlabs.node.api.casper._
 import io.casperlabs.shared.Log
-import io.casperlabs.comm.ServiceError.InvalidArgument
 import io.casperlabs.smartcontracts.ExecutionEngineService
-import io.casperlabs.models.SmartContractEngineError
-import io.casperlabs.ipc
-import monix.execution.Scheduler
 import monix.eval.{Task, TaskLike}
 import monix.reactive.Observable
 
@@ -124,7 +123,7 @@ object GrpcCasperService extends StateConversions {
     }
 
   def toKey[F[_]: MonadThrowable](keyType: StateQuery.KeyVariant, keyValue: String): F[ipc.Key] =
-    GrpcDeployService.toKey[F](keyType.name, keyValue).handleErrorWith {
+    Utils.toKey[F](keyType.name, keyValue).handleErrorWith {
       case ex: java.lang.IllegalArgumentException =>
         MonadThrowable[F].raiseError(InvalidArgument(ex.getMessage))
     }
